Store weight as DECIMAL and fix conversion in display

+++ db/init_db.sql
@@ -4,7 +4,7 @@ CREATE TABLE IF NOT EXISTS travel_items (
     id INT AUTO_INCREMENT PRIMARY KEY,
     -- UNIQUE on name is paired with checks in the front-end to make sure names are unique --
     name VARCHAR(255) UNIQUE,
-    weight INT,
+    weight DECIMAL(20, 3),
     created_at DATETIME DEFAULT CURRENT_TIMESTAMP,
     updated_at DATETIME DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP
 );
